Net5 0.6w

+++ InSimDotNet/Packets/ISendable.cs
@@ -8,7 +8,7 @@ namespace InSimDotNet.Packets {
         /// <summary>
         /// Gets the size of the packet.
         /// </summary>
-        byte Size { get; }
+        int Size { get; }
 
         /// <summary>
         /// Gets the type of the packet.
